fixed submit of nominee forms, and added lookup fields

+++ db/migrate/20130108225011_add_basics_to_nominees.rb
@@ -2,7 +2,7 @@ class AddBasicsToNominees < ActiveRecord::Migration
   def change
     add_column :nominees, :age, :integer
     add_column :nominees, :city, :string
-    add_column :nominees, :school_grade, :integer
+    add_column :nominees, :school_grade, :decimal
     add_column :nominees, :math_units, :integer
     add_column :nominees, :grade, :integer
     add_column :nominees, :sat_grade, :integer
